FLUME-263: DirWatcher does not match filename of deleted files against supplied REGEX
If you supply a regex to TailDirSource then it also supplies the regex to DirWatcher, so that DirWatcher only informs TailDirSource about the files that match the regex.
Now when files are deleted from a directory, DirWatcher will match the file's name against the supplied regex and only a match occurs it will fire an event to TailDirSource about the deleted file.